Disable all regext matches with rmt

+++ src/commands/custom/favna.js
@@ -26,6 +26,7 @@
 /**
  * @file Custom FavnaCommand - Custom Command exclusive to ChaosGamez server  
  * A joke command to praise Favna  
+ * Server admins can disable this command entirely by using the `rmt off` command  
  * **Aliases**: `.favna`
  * @author Jeroen Claassens (favna) <sharkie.jeroen@gmail.com>
  * @module
@@ -58,7 +59,8 @@ module.exports = class FavnaCommand extends commando.Command {
     if (this.client.isOwner(msg.author)) {
       return true;
     }
-    if (msg.guild.id !== '373826006651240450') {
+
+    if (msg.guild.id !== '373826006651240450' && this.client.provider.get(msg.guild.id, 'regexmatches', false)) {
       return stripIndents`That command can only be used in the Chaos Gamez server, sorry 😦
 			Want your own server specific custom commands? Join the support server (link in the \`${msg.guild.commandPrefix}stats\` command) and request the command.`;
     }
@@ -67,14 +69,16 @@ module.exports = class FavnaCommand extends commando.Command {
   }
 
   run (msg) {
-    msg.delete();
-    msg.embed({
-      'image': {'url': 'https://favna.xyz/images/ribbonhost/favnadedsec.gif'},
-      'color': msg.guild ? msg.guild.me.displayColor : 10610610,
-      'description': oneLine`Technically speaking my father, but to you he is your supreme leader and you will submit to him 
-            or I will infect every single human being you have ever met with a virus so terrible their lungs and intestines
-             will instantly explode from their chests causing a gorey, bloody mess all over the floor and you 
-             will be the only person held responsible for the death of hundredths if not millions of people.`
-    });
+    if (this.client.provider.get(msg.guild.id, 'regexmatches', false)) {
+      msg.delete();
+      msg.embed({
+        'image': {'url': 'https://favna.xyz/images/ribbonhost/favnadedsec.gif'},
+        'color': msg.guild ? msg.guild.me.displayColor : 10610610,
+        'description': oneLine`Technically speaking my father, but to you he is your supreme leader and you will submit to him 
+          or I will infect every single human being you have ever met with a virus so terrible their lungs and intestines
+           will instantly explode from their chests causing a gorey, bloody mess all over the floor and you 
+           will be the only person held responsible for the death of hundredths if not millions of people.`
+      });
+    }
   }
 };
