+++ src/commands/custom/kai.js
@@ -26,6 +26,7 @@
 /**
  * @file Custom KaiCommand - Custom Command exclusive to ChaosGamez server  
  * A joke command to spite Kai  
+ * Server admins can disable this command entirely by using the `rmt off` command  
  * **Aliases**: `.kai`
  * @author Jeroen Claassens (favna) <sharkie.jeroen@gmail.com>
  * @module
@@ -74,7 +75,8 @@ module.exports = class KaiCommand extends commando.Command {
     if (this.client.isOwner(msg.author)) {
       return true;
     }
-    if (msg.guild.id !== '373826006651240450') {
+
+    if (msg.guild.id !== '373826006651240450' && this.client.provider.get(msg.guild.id, 'regexmatches', false)) {
       return stripIndents`That command can only be used in the Chaos Gamez server, sorry 😦
 			Want your own server specific custom commands? Join the support server (link in the \`${msg.guild.commandPrefix}stats\` command) and request the command.`;
     }
@@ -83,11 +85,13 @@ module.exports = class KaiCommand extends commando.Command {
   }
 
   run (msg) {
-    msg.delete();
-    msg.embed({
-      'image': {'url': this.fetchImage()},
-      'color': msg.guild ? msg.guild.me.displayColor : 10610610
-    },
-    'Please <@418504046337589249> get lost');
+    if (this.client.provider.get(msg.guild.id, 'regexmatches', false)) {
+      msg.delete();
+      msg.embed({
+        'image': {'url': this.fetchImage()},
+        'color': msg.guild ? msg.guild.me.displayColor : 10610610
+      },
+      'Please <@418504046337589249> get lost');
+    }
   }
 };
